# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "效率仪表盘后端API，提供任务、提交、用户、仓库、项目、组织等数据管理接口",
        "title": "Efficiency Dashboard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9990",
    "basePath": "/api",
    "paths": {
...
        "/api/v2/dashboard/summary": {
            "get": {
                "description": "获取仪表盘的汇总统计数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "获取仪表盘汇总信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.DashboardSummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.DashboardSummaryResponse": {
            "type": "object",
            "properties": {
                "avg_efficiency_ratio": {
                    "type": "number"
                },
                "total_commits": {
                    "type": "integer"
                },
                "total_cost": {
                    "type": "number"
                },
                "total_diff_lines": {
                    "type": "integer"
                },
                "total_real_minutes": {
                    "type": "number"
                },
                "total_repos": {
                    "type": "integer"
                },
                "total_task_ancient_minutes": {
                    "type": "number"
                },
                "total_tasks": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                },
                "total_work_dirs": {
                    "type": "integer"
                }
            }
        },
...
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
    }
}